Clarified scale requirement warnings...
Addresses issues in #14

Specifically:
1. GOF Test Statistic rounded to 2 digits
2. PVal rounded to 4 digits
3. model.type = "ssm" is the default

New type: "auto"
Automatically detects and assigns whether we should run additional features on the data.

"auto" is applied to inference and compute.v

Samples > 10,000, default to "fast"

Otherwise, we use inference and compute.v

Misc:
inference is auto-disabled if ARMA is detected until we can create the matrix.

Former-commit-id: 2f3a215cc20eecf0da5f07c507ae50518aef799a